Refactored schedule to move non-views to utilities
