+++ sd_food_bank_ai_bot/admin_panel/views/utilities.py
@@ -5,7 +5,13 @@ from twilio.rest import Client
 from django.conf import settings
 from openai import OpenAI
 from django.views.decorators.csrf import csrf_exempt
+from django.utils.timezone import now
+from datetime import time, datetime, timedelta
 import re
+
+EARLIEST_TIME = time(9, 0)   # Earliest time to schedule an appointment, 9:00 AM
+LATEST_TIME = time(17, 0)    # Latest time appointments can end, 5:00 PM
+FIXED_APPT_DURATION = timedelta(minutes=15) # TODO: Assuming each appointment is 15 minutes
 
 def strike_system_handler(log, reset = False):
     """Updates strikes within the log object associated with call as conversation progresses"""
@@ -31,7 +37,7 @@ def get_phone_number(request):
         return caller_number
     return None
 
-def get_response_sentiment(request, sentence):
+def get_response_sentiment(sentence):
     """
     Returns True if the given sentence is affirmative
     """
@@ -57,7 +63,7 @@ def return_main_menu(request):
     Redirects user to main menu based on YES or NO sentiment
     """
     speech_result = request.POST.get('SpeechResult', '').strip().lower()
-    declaration = get_response_sentiment(request, speech_result)
+    declaration = get_response_sentiment(speech_result)
     response = VoiceResponse()
 
     if declaration:
@@ -166,3 +172,90 @@ def get_corresponding_answer(question):
     """
     answer = FAQ.objects.filter(question__iexact=question).first().answer
     return answer
+
+def get_day(speech_result):
+    """
+    Extracts the day of the week from a given message. Only returns the day or NONE.
+    """
+    client = OpenAI()
+    system_prompt = "Please extract the day of the week from the following message. Only respond with the day of the week or NONE if one is not said."
+    completion = client.chat.completions.create(
+        model="gpt-4o-mini",
+        messages=[
+            {"role": "system", "content": system_prompt},
+            {"role": "user", "content": speech_result}
+        ]
+    )   
+    response_pred = completion.choices[0].message.content
+    
+    return response_pred
+
+def check_available_date(target_weekday):
+    """
+    Return if the given date has available timeslots or not.
+    """
+    today = now().date()
+    max_weeks_ahead = 4 
+    number_available_appointments = 0
+
+    for week in range(max_weeks_ahead):
+        # Calculate the next occurrence of the requested weekday
+        days_ahead = (target_weekday - today.weekday()) % 7
+        if days_ahead == 0 and week == 0:
+            days_ahead = 7 
+
+        appointment_date = today + timedelta(days=days_ahead + (week * 7))
+
+        # Retrieve existing appointments for this date
+        existing_appointments = AppointmentTable.objects.filter(date__date=appointment_date).order_by('start_time')
+
+        # For no appointments that day yet
+        if not existing_appointments:
+            number_available_appointments = 4 # TODO: mod by n = fixed appt. length
+            return True, appointment_date, number_available_appointments
+
+        current_time = EARLIEST_TIME
+
+        # Check for time slots before appointments until the latest appointment
+        for appointment in existing_appointments:
+            while current_time < appointment.start_time:
+                number_available_appointments += 1
+                # Increment by fixed appt time
+                current_time = (datetime.combine(datetime.today(), current_time) + FIXED_APPT_DURATION).time()
+            current_time = appointment.end_time
+
+        # Checks for time slots after last appointment is iterated in the for loop above
+        while current_time < LATEST_TIME:
+            number_available_appointments += 1
+            current_time = (datetime.combine(datetime.today(), current_time) + FIXED_APPT_DURATION).time()
+        
+        # If there are available timeslots return True and additional var.
+        if number_available_appointments > 0:
+            return True, appointment_date, number_available_appointments
+
+    # If no available timeslots for the next month on request day return False    
+    return False, None, 0
+
+def get_available_times_for_date(appointment_date):
+    """
+    Retrieve available appointment times for a given date.
+    """
+    existing_appointments = AppointmentTable.objects.filter(date__date=appointment_date).order_by('start_time')
+    available_times = []
+
+    current_time = EARLIEST_TIME
+    
+    # Adds timeslots between appointments
+    for appointment in existing_appointments:
+        if current_time < appointment.start_time:
+            available_times.append(current_time)
+            # Increment by fixed appt time
+            current_time = (datetime.combine(datetime.today(), current_time) + FIXED_APPT_DURATION).time()
+        current_time = appointment.end_time
+
+    # Adds timeslots after the latest iterated appointment
+    while current_time < LATEST_TIME:
+        available_times.append(current_time)
+        current_time = (datetime.combine(datetime.today(), current_time) + FIXED_APPT_DURATION).time()
+
+    return available_times
